+++ src/main/java/org/opentripplanner/routing/service/DefaultRoutingService.java
@@ -28,6 +28,7 @@ public class DefaultRoutingService implements RoutingService {
 
   @Override
   public RoutingResponse route(RouteRequest request) {
+    request.validateOriginAndDestination();
     RoutingWorker worker = new RoutingWorker(serverContext, request, timeZone);
     return worker.route();
   }